web/requirements.txt
django
pillow
django-formtools
django-rq
pyserial
redis
